trivial refactor
